inject go modules directly for easier debugging.
